Getting rid of dual versions of the converter, tests are in, native is faster
